import_map in deno.json, isolate test deps, add constants file
