+++ src/deps.ts
@@ -10,15 +10,6 @@ export { platform } from "https://deno.land/std@0.173.0/node/os.ts";
 export { walk } from "https://deno.land/std@0.180.0/fs/mod.ts";
 export * as JSONC from "https://deno.land/std@0.180.0/encoding/jsonc.ts";
 export { delay } from "https://deno.land/std@0.180.0/async/delay.ts";
-//  - testing
-export {
-  afterAll,
-  beforeAll,
-  beforeEach,
-  describe,
-  it,
-} from "https://deno.land/std@0.180.0/testing/bdd.ts";
-export { assert } from "https://deno.land/std@0.180.0/testing/asserts.ts";
 
 // Third party
 //  - cliffy
@@ -78,23 +69,3 @@ export const ccolors = {
     return `${colors.white(errCode)}: ${colors.red(e.toString())}`;
   },
 };
-
-// constants
-export const TERRAFORM_VERSION = "1.4.6";
-export const KUBESEAL_VERSION = "0.19.1";
-
-export const DEFAULT_INSTANCE_TYPES = {
-  aws: "m5a.large" as const,
-  gcp: "n2-standard-2" as const,
-  azure: "Standard_D4s_v3" as const,
-};
-
-export const DEFAULT_NODE_DISK_SIZE = 100;
-
-export const NODE_DISK_SIZE_KEY = {
-  aws: "volume_size" as const,
-  gcp: "size" as const,
-  azure: "disk_size_gb" as const,
-};
-
-export const nonMicrok8sNodeKinds = ["eks"];
