+++ src/outputs/terraform/aws-eks/cndi_aws_eks_node_group.tf.json.ts
@@ -1,5 +1,5 @@
 import { getPrettyJSONString, getTFResource } from "src/utils.ts";
-import { DEFAULT_INSTANCE_TYPES, DEFAULT_NODE_DISK_SIZE } from "deps";
+import { DEFAULT_INSTANCE_TYPES, DEFAULT_NODE_DISK_SIZE } from "constants";
 import { AWSEKSNodeItemSpec } from "src/types.ts";
 
 export default function getAWSEKServiceNodeGroupTFJSON(
