+++ src/tests/deps.ts
@@ -1,0 +1,7 @@
+export {
+  beforeEach,
+  describe,
+  it,
+} from "https://deno.land/std@0.180.0/testing/bdd.ts";
+
+export { assert } from "https://deno.land/std@0.180.0/testing/asserts.ts";
