+++ src/tests/helpers/util.ts
@@ -1,4 +1,5 @@
-import { assert, path } from "deps";
+import { assert } from "test-deps";
+import { path } from "deps";
 
 function areSetsEqual(setA: Set<string>, setB: Set<string>) {
   if (setA.size !== setB.size) return false;
